Feat: Organize files by directories
