+++ components/Favorites.jsx
