+++ index.html
@@ -5,7 +5,7 @@
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Document</title>
-	<link href="./styles.css" type="text/css" rel="stylesheet" />
+	<link href="./styles/styles.css" type="text/css" rel="stylesheet" />
 </head>
 
 <body>
