+++ pages/App.jsx
@@ -91,7 +91,7 @@ const App = () => {
 
     function getFavoritesFromLocalStorage() {
         const favorites = localStorage.getItem('favorites');
-        if (favorites || favorites.length) {
+        if (favorites && favorites.length) {
             setFavorites(JSON.parse(favorites));
         }
     }
